N/A

+++ src/main/java/com/yasser/InventoryTrack/controller/ProductController.java
@@ -2,8 +2,10 @@ package com.yasser.InventoryTrack.controller;
 
 import com.yasser.InventoryTrack.dto.ProductsDTO;
 import com.yasser.InventoryTrack.service.ServiceRepository;
+import com.yasser.InventoryTrack.util.SecurityContext;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -14,10 +16,12 @@ import java.util.List;
 public class ProductController {
 
     private final ServiceRepository serviceRepository;
+    private final SecurityContext securityContext;
 
-    public ProductController(ServiceRepository serviceRepository) {
+    public ProductController(ServiceRepository serviceRepository, SecurityContext securityContext) {
 
         this.serviceRepository = serviceRepository;
+        this.securityContext = securityContext;
     }
 
 
@@ -30,6 +34,7 @@ public class ProductController {
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
     public List<ProductsDTO> getAllProduct() {
+
         return serviceRepository.getProducts();
     }
 
@@ -40,9 +45,14 @@ public class ProductController {
     }
 
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteProduct(@PathVariable int id) throws ResponseStatusException{
-        serviceRepository.deleteProduct(id);
+    public ResponseEntity<?> deleteProduct(@PathVariable int id) throws ResponseStatusException{
+        List<String> roles = this.securityContext.getRoles();
+        if (roles.contains("owner")) {
+            serviceRepository.deleteProduct(id);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
+
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
 }
